задание 3
