+++ homework.txt
@@ -7,6 +7,9 @@
 Framework/Registry. Реестр позволяет получить доступ к парам ключ-объект из любого места приложения. Здесь мы получаем данные из конфигурационного файла (getDataConfig), и храним метод рендера страницы по названию роута.
 
 3. Разобраться и объяснить, какой тип модели используется в проекте.
-
+По идее, доменная модель, особенно если учитываем все наши добавления паттернов по домашкам. У нас есть слои объектов, 
+* в самом дальнем от ядра – пользовательский клиент|на стороне инфраструктуры - рассылка сообщений, репосты в соц сети. 
+* дальше уровень с командами, и контроллерами, и сервисы| адаптары в нашим соц рассылкам и тп. 
+* домен, наше приложение. Модели 
 
 4. Переписать проект, используя паттерн Page Controller.
