+++ src/Controller/ProductInfoController.php
@@ -1,16 +1,7 @@
 <?php
-
-declare(strict_types=1);
-
 namespace Controller;
 
-use Framework\BaseController;
-use Service\Order\Basket;
-use Service\Product\ProductService;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-
-class ProductController extends BaseController
+class ProductInfoController extends BaseController
 {
     /**
      * Информация о продукте
@@ -37,23 +28,6 @@ class ProductController extends BaseController
         return $this->render(
             'product/info.html.php',
             ['productInfo' => $productInfo, 'isInBasket' => $isInBasket]
-        );
-    }
-
-    /**
-     * Список всех продуктов
-     * @param Request $request
-     * @return Response
-     */
-    public function listAction(Request $request): Response
-    {
-        $productList = (new ProductService())->getAll(
-            $request->query->get('sort', '')
-        );
-
-        return $this->render(
-            'product/list.html.php',
-            ['productList' => $productList]
         );
     }
 }
